shelving this branch
krh thought the task of repackaging retrace functions as a libray was
unnecessary at this point.

He feels like this refactoring should be done at a later date, once
the value of the project is established.  He sees no issue with
forking apitrace for now, and using git to pull changes as needed.